Downgrade to exoplayer 1.0.2

+++ cobalt/android/apk/app/src/main/java/dev/cobalt/media/ExoPlayerBridge.java
@@ -402,7 +402,7 @@ public class ExoPlayerBridge {
   }
 
   @UsedByNative
-  private synchronized int getDroppedFrames() {
+  private synchronized long getDroppedFrames() {
     return droppedFramesOnPlayerThread;
   }
